# openapi.yaml
openapi: 3.0.3
info:
  title: Hypernova Banking API
  version: 1.0.0
  license:
    name: Proprietary technical assessment implementation
    url: https://github.com/0xStyx507/hypernova-banking
  description: |
    Contrato HTTP de Hypernova Banking.

    Incluye identidad, sesiones, auditoría, cuentas USD y movimientos
    financieros respaldados por TigerBeetle. Los importes son unidades menores
    enteras serializadas como cadenas para evitar redondeos de punto flotante.
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
# Public health and authentication operations opt in to security below; an
# empty root requirement makes the security posture explicit for every path.
security: []
...
paths:
...
  /api/v1/auth/mfa/enroll:
    post:
      operationId: enrollMFA
      summary: Genera un enrolamiento TOTP de un solo uso
      description: Devuelve una URI otpauth estándar para crear un QR compatible con Google Authenticator y Microsoft Authenticator. El secreto solo se devuelve en esta respuesta.
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: Enrolamiento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFAEnrollment'
        '401':
          description: Sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '503':
          description: MFA no disponible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/auth/mfa/verify:
    post:
      operationId: verifyMFA
      summary: Activa el MFA tras validar un código TOTP
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFACodeRequest'
      responses:
        '200':
          description: MFA activado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFAStatus'
        '401':
          description: Código inválido o sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: Enrolamiento ausente o MFA ya activo
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '410':
          description: Enrolamiento expirado
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/auth/mcp-pin:
    get:
      operationId: getMCPPINStatus
      summary: Consulta si el usuario tiene un PIN MCP configurado
      description: Devuelve el estado y vencimiento del PIN; nunca expone el hash ni el PIN.
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Estado del PIN MCP
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPPINStatus' } } }
        '401':
          description: Sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    post:
      operationId: setMCPPIN
      summary: Crea o actualiza el PIN de confirmación MCP
      description: Requiere una sesión válida con MFA verificado. El PIN debe contener exactamente cuatro dígitos ASCII, se almacena como hash bcrypt y vence en tres minutos; al vencer se limpia de PostgreSQL.
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MCPPINRequest' } } }
      responses:
        '200':
          description: PIN configurado
          content: { application/json: { schema: { $ref: '#/components/schemas/MCPPINStatus' } } }
        '400':
          description: El PIN no contiene exactamente cuatro dígitos
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Sesión inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '403':
          description: MFA no habilitado o no verificado en la sesión actual
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '503':
          description: Almacén de identidad no disponible
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/accounts:
    post:
      operationId: createAccount
      summary: Crea otra cuenta USD para el usuario autenticado
      description: Cada cuenta pertenece al usuario autenticado y puede seleccionarse como origen de una operación. La identidad recibe una cuenta inicial durante el registro.
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Cuenta provisionada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Solicitud inválida
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    get:
      operationId: listAccounts
      summary: Lista las cuentas propias
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Cuentas del usuario
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
        '401':
          description: Token inválido o ausente
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Mensaje público y no sensible del error
          example: invalid credentials
        code:
          type: string
          description: Código estable para clientes
          example: insufficient_funds
...
    MFACodeRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          pattern: '^[0-9]{6}$'
    MFAStatus:
      type: object
      required: [enabled, enrolled]
      properties:
        enabled: { type: boolean }
        enrolled: { type: boolean }
    MFAEnrollment:
      type: object
      required: [secret, otpauth_uri, expires_at]
      properties:
        secret:
          type: string
          description: Secreto TOTP mostrado una sola vez para recuperación manual.
        otpauth_uri:
          type: string
          format: uri
          description: URI estándar para generar un QR compatible con autenticadores TOTP.
        expires_at: { type: string, format: date-time }
...
    CreateAccountRequest:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
          enum: [USD]
          default: USD
...
    Account:
      type: object
      required: [id, display_name, currency, type, status, created_at]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string, maxLength: 48 }
        currency: { type: string, enum: [USD] }
        type: { type: string, enum: [checking] }
        status: { type: string, enum: [provisioning, active, failed, closed] }
        created_at: { type: string, format: date-time }
...
    MCPPINRequest:
      type: object
      additionalProperties: false
      required: [pin]
      properties:
        pin: { type: string, pattern: '^[0-9]{4}$', description: PIN MCP de exactamente cuatro dígitos ASCII }
    MCPPINStatus:
      type: object
      required: [configured]
      properties:
        configured: { type: boolean, description: Indica si existe un PIN vigente para el usuario }
        expires_at: { type: string, format: date-time, description: Hora de vencimiento del PIN vigente }
...